requirements.txt
aiogram==3.13.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
apscheduler==3.10.4
openpyxl==3.1.5
reportlab==4.2.5
python-dotenv==1.0.1
redis==5.2.0
aioredis==2.0.1
Pillow==11.0.0
pytz==2024.2
